# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Logoviking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/youtube-earnings:
    post:
      operationId: calcYoutubeEarnings
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YoutubeEarningsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsResult"
  /tools/tiktok-earnings:
    post:
      operationId: calcTiktokEarnings
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TiktokEarningsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsResult"
  /tools/instagram-earnings:
    post:
      operationId: calcInstagramEarnings
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstagramEarningsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsResult"
  /tools/engagement-rate:
    post:
      operationId: calcEngagementRate
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EngagementRateInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementRateResult"
  /tools/thumbnail-downloader:
    post:
      operationId: downloadThumbnail
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThumbnailDownloadInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThumbnailDownloadResult"
  /tools/channel-name-generator:
    post:
      operationId: generateChannelName
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringListResult"
  /tools/hashtag-generator:
    post:
      operationId: generateHashtags
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashtagInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringListResult"
  /tools/caption-generator:
    post:
      operationId: generateCaption
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringListResult"
  /tools/youtube-title-generator:
    post:
      operationId: generateYoutubeTitle
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringListResult"
  /tools/content-idea-generator:
    post:
      operationId: generateContentIdeas
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringListResult"
  /tools/thumbnail-text-generator:
    post:
      operationId: generateThumbnailText
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringListResult"
  /tools/ai-thumbnail-generator:
    post:
      operationId: generateAiThumbnail
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiThumbnailResult"
  /tools/meta-tag-generator:
    post:
      operationId: generateMetaTags
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetaTagInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaTagResult"
  /tools/keyword-density:
    post:
      operationId: keywordDensity
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordDensityInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordDensityResult"
  /tools/seo-title-generator:
    post:
      operationId: generateSeoTitle
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringListResult"
  /tools/sitemap-generator:
    post:
      operationId: generateSitemap
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SitemapInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringResult"
  /tools/robots-txt-generator:
    post:
      operationId: generateRobotsTxt
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RobotsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringResult"
...
  /tools/creator-kit:
    post:
      operationId: generateCreatorKit
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorKitResult"
...
components:
  schemas:
...
    StringResult:
      type: object
      properties:
        value: { type: string }
      required: [value]
    StringListResult:
      type: object
      properties:
        items:
          type: array
          items: { type: string }
      required: [items]
    YoutubeEarningsInput:
      type: object
      properties:
        views: { type: number }
        cpm: { type: number }
        engagementRate: { type: number }
      required: [views]
    TiktokEarningsInput:
      type: object
      properties:
        views: { type: number }
        followers: { type: number }
      required: [views]
    InstagramEarningsInput:
      type: object
      properties:
        followers: { type: number }
        engagementRate: { type: number }
      required: [followers]
    EarningsResult:
      type: object
      properties:
        low: { type: number }
        avg: { type: number }
        high: { type: number }
        currency: { type: string }
        breakdown:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              value: { type: string }
            required: [label, value]
      required: [low, avg, high, currency]
    EngagementRateInput:
      type: object
      properties:
        followers: { type: number }
        likes: { type: number }
        comments: { type: number }
        shares: { type: number }
      required: [followers, likes]
    EngagementRateResult:
      type: object
      properties:
        rate: { type: number }
        rating: { type: string }
        tip: { type: string }
      required: [rate, rating]
    ThumbnailDownloadInput:
      type: object
      properties:
        url: { type: string }
      required: [url]
    ThumbnailDownloadResult:
      type: object
      properties:
        videoId: { type: string }
        thumbnails:
          type: array
          items:
            type: object
            properties:
              quality: { type: string }
              url: { type: string }
            required: [quality, url]
      required: [videoId, thumbnails]
    KeywordInput:
      type: object
      properties:
        keyword: { type: string }
      required: [keyword]
    HashtagInput:
      type: object
      properties:
        keyword: { type: string }
        platform: { type: string }
      required: [keyword]
    CaptionInput:
      type: object
      properties:
        keyword: { type: string }
        tone: { type: string }
      required: [keyword]
    AiThumbnailResult:
      type: object
      properties:
        layouts:
          type: array
          items:
            type: object
            properties:
              title: { type: string }
              text: { type: string }
              colors:
                type: array
                items: { type: string }
              tip: { type: string }
            required: [title, text, colors, tip]
      required: [layouts]
    MetaTagInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        url: { type: string }
        image: { type: string }
        keywords: { type: string }
      required: [title, description]
    MetaTagResult:
      type: object
      properties:
        html: { type: string }
      required: [html]
    KeywordDensityInput:
      type: object
      properties:
        text: { type: string }
      required: [text]
    KeywordDensityResult:
      type: object
      properties:
        wordCount: { type: number }
        keywords:
          type: array
          items:
            type: object
            properties:
              word: { type: string }
              count: { type: number }
              density: { type: number }
            required: [word, count, density]
      required: [wordCount, keywords]
    SitemapInput:
      type: object
      properties:
        urls:
          type: array
          items: { type: string }
      required: [urls]
    RobotsInput:
      type: object
      properties:
        allowAll: { type: boolean }
        sitemapUrl: { type: string }
        disallowed:
          type: array
          items: { type: string }
      required: [allowAll]
...
    CreatorKitResult:
      type: object
      properties:
        thumbnailIdea: { type: string }
        title: { type: string }
        hashtags:
          type: array
          items: { type: string }
        caption: { type: string }
      required: [thumbnailIdea, title, hashtags, caption]
...