# lib/api-zod/src/generated/types/aiThumbnailResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logoviking API
 * OpenAPI spec version: 0.1.0
 */
import type { AiThumbnailResultLayoutsItem } from "./aiThumbnailResultLayoutsItem";

export interface AiThumbnailResult {
  layouts: AiThumbnailResultLayoutsItem[];
}