# lib/api-zod/src/generated/types/engagementRateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logoviking API
 * OpenAPI spec version: 0.1.0
 */

export interface EngagementRateResult {
  rate: number;
  rating: string;
  tip?: string;
}