# lib/api-zod/src/generated/types/keywordDensityResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logoviking API
 * OpenAPI spec version: 0.1.0
 */
import type { KeywordDensityResultKeywordsItem } from "./keywordDensityResultKeywordsItem";

export interface KeywordDensityResult {
  wordCount: number;
  keywords: KeywordDensityResultKeywordsItem[];
}