# lib/api-zod/src/generated/types/metaTagInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logoviking API
 * OpenAPI spec version: 0.1.0
 */

export interface MetaTagInput {
  title: string;
  description: string;
  url?: string;
  image?: string;
  keywords?: string;
}