# lib/api-zod/src/generated/types/thumbnailDownloadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logoviking API
 * OpenAPI spec version: 0.1.0
 */
import type { ThumbnailDownloadResultThumbnailsItem } from "./thumbnailDownloadResultThumbnailsItem";

export interface ThumbnailDownloadResult {
  videoId: string;
  thumbnails: ThumbnailDownloadResultThumbnailsItem[];
}